Enormous Grinding
After many hours of grinding with a net that didn't work, i believe it now works, at least fundamentally.

main issue was unlinked CNNs; the internal feature extractor CNN was not linked to the main CNN's model, and was being re-initialized on every forward() call. gah. Goes to prove the point that small iterative changes as-you-build (especially in domains with poor familiarity) is extremely important.